menambah css untuk memperindah tampilan lalu mengintegrasikannya ke file php

+++ delete.php
@@ -1,10 +1,12 @@
 <?php
 require 'connection.php';
-$id = $_GET['id'];
-$sql = "DELETE FROM mahasiswa WHERE id = ?";
-$stmt = $pdo->prepare($sql);
-if ($stmt->execute([$id])) {
+
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
+    $sql = "DELETE FROM mahasiswa WHERE id = ?";
+    $stmt = $pdo->prepare($sql);
+    $stmt->execute([$id]);
+}
+
 header("Location: index.php");
 exit;
-}
-?>
